process-monitor/requirements.txt
psutil==7.0.0
pystray==0.19.5
Pillow==11.2.1